Factor out GridColumn to its own model file to prevent circular dependencies in jest test
